:ok_hand: PR feedback
* Update CLI documentation
* Clean up notifications test mixin(s) into a single mixin
