+++ src/openzaak/components/besluiten/tests/test_notifications_send.py
@@ -16,12 +16,9 @@ from openzaak.components.documenten.tests.factories import (
     EnkelvoudigInformatieObjectFactory,
 )
 from openzaak.notifications.models import FailedNotification
-from openzaak.notifications.tests import mock_nrc_oas_get
-from openzaak.notifications.tests.utils import (
-    LOGGING_SETTINGS,
-    NotificationsConfigMixin,
-    get_notifications_api_root,
-)
+from openzaak.notifications.tests import mock_notification_send, mock_nrc_oas_get
+from openzaak.notifications.tests.mixins import NotificationsConfigMixin
+from openzaak.notifications.tests.utils import LOGGING_SETTINGS
 from openzaak.tests.utils import JWTAuthMixin
 
 from ..constants import VervalRedenen
@@ -35,12 +32,6 @@ from .utils import get_operation_url
 class SendNotifTestCase(NotificationsConfigMixin, JWTAuthMixin, APITestCase):
 
     heeft_alle_autorisaties = True
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-
-        cls._configure_notifications()
 
     @patch("zds_client.Client.from_url")
     def test_send_notif_create_besluit(self, m, mock_client):
@@ -131,15 +122,9 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
     heeft_alle_autorisaties = True
     maxDiff = None
 
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-
-        cls._configure_notifications()
-
     def test_besluit_create_fail_send_notification_create_db_entry(self, m):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         besluittype = BesluitTypeFactory.create(concept=False)
         besluittype_url = reverse(besluittype)
         url = get_operation_url("besluit_create")
@@ -183,7 +168,7 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
 
     def test_besluit_delete_fail_send_notification_create_db_entry(self, m):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         besluit = BesluitFactory.create()
         url = reverse(besluit)
 
@@ -216,7 +201,7 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
         self, m
     ):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         url = get_operation_url("besluitinformatieobject_create")
 
         besluit = BesluitFactory.create()
@@ -262,7 +247,7 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
         self, m
     ):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         bio = BesluitInformatieObjectFactory.create()
         url = reverse(bio)
