+++ src/openzaak/components/catalogi/tests/admin/test_admin_publish.py
@@ -9,7 +9,7 @@ import requests_mock
 from django_webtest import WebTest
 
 from openzaak.accounts.tests.factories import SuperUserFactory, UserFactory
-from openzaak.notifications.tests.utils import NotificationsConfigMixin
+from openzaak.notifications.tests.mixins import NotificationsConfigMixin
 from openzaak.selectielijst.models import ReferentieLijstConfig
 from openzaak.selectielijst.tests import (
     mock_resource_get,
@@ -53,7 +53,6 @@ class ZaaktypeAdminTests(
 
     @override_settings(NOTIFICATIONS_DISABLED=False)
     def test_publish_zaaktype(self, m):
-        self._configure_notifications()
         procestype_url = (
             "https://selectielijst.openzaak.nl/api/v1/"
             "procestypen/e1b73b12-b2f6-4c4e-8929-94f84dd2a57d"
