+++ src/openzaak/components/catalogi/tests/admin/test_notifications.py
@@ -7,7 +7,7 @@ import requests_mock
 from django_webtest import WebTest
 
 from openzaak.accounts.tests.factories import SuperUserFactory
-from openzaak.notifications.tests.utils import NotificationsConfigMixin
+from openzaak.notifications.tests.mixins import NotificationsConfigMixin
 from openzaak.selectielijst.models import ReferentieLijstConfig
 from openzaak.selectielijst.tests import (
     mock_resource_get,
@@ -40,8 +40,6 @@ class NotificationAdminTests(
         config = ReferentieLijstConfig.get_solo()
         config.allowed_years = [2017, 2020]
         config.save()
-
-        cls._configure_notifications()
 
         cls.user = SuperUserFactory.create()
