+++ src/openzaak/components/documenten/tests/test_notifications_send.py
@@ -17,12 +17,9 @@ from vng_api_common.tests import reverse
 from openzaak.components.catalogi.tests.factories import InformatieObjectTypeFactory
 from openzaak.components.documenten.models import EnkelvoudigInformatieObject
 from openzaak.notifications.models import FailedNotification
-from openzaak.notifications.tests import mock_nrc_oas_get
-from openzaak.notifications.tests.utils import (
-    LOGGING_SETTINGS,
-    NotificationsConfigMixin,
-    get_notifications_api_root,
-)
+from openzaak.notifications.tests import mock_notification_send, mock_nrc_oas_get
+from openzaak.notifications.tests.mixins import NotificationsConfigMixin
+from openzaak.notifications.tests.utils import LOGGING_SETTINGS
 from openzaak.tests.utils import JWTAuthMixin
 
 from .factories import EnkelvoudigInformatieObjectFactory, GebruiksrechtenFactory
@@ -35,12 +32,6 @@ from .utils import get_operation_url
 class SendNotifTestCase(NotificationsConfigMixin, JWTAuthMixin, APITestCase):
 
     heeft_alle_autorisaties = True
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-
-        cls._configure_notifications()
 
     @patch("zds_client.Client.from_url")
     def test_send_notif_create_enkelvoudiginformatieobject(self, m, mock_client):
@@ -96,15 +87,9 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
     heeft_alle_autorisaties = True
     maxDiff = None
 
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-
-        cls._configure_notifications()
-
     def test_eio_create_fail_send_notification_create_db_entry(self, m):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         url = get_operation_url("enkelvoudiginformatieobject_create")
 
         informatieobjecttype = InformatieObjectTypeFactory.create(concept=False)
@@ -156,7 +141,7 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
 
     def test_eio_delete_fail_send_notification_create_db_entry(self, m):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         eio = EnkelvoudigInformatieObjectFactory.create()
         url = reverse(eio)
 
@@ -188,7 +173,7 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
 
     def test_gebruiksrechten_create_fail_send_notification_create_db_entry(self, m):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         url = get_operation_url("gebruiksrechten_create")
 
         eio = EnkelvoudigInformatieObjectFactory.create()
@@ -229,7 +214,7 @@ class FailedNotificationTests(NotificationsConfigMixin, JWTAuthMixin, APITestCas
 
     def test_gebruiksrechten_delete_fail_send_notification_create_db_entry(self, m):
         mock_nrc_oas_get(m)
-        m.post(f"{get_notifications_api_root()}notificaties", status_code=403)
+        mock_notification_send(m, status_code=403)
         gebruiksrechten = GebruiksrechtenFactory.create()
 
         url = reverse(gebruiksrechten)
