+++ src/openzaak/notifications/tests/__init__.py
@@ -8,10 +8,12 @@ def _get_base_url() -> str:
     from notifications_api_common.models import NotificationsConfig
 
     config = NotificationsConfig.get_solo()
-    base_url = (
-        getattr(config.notifications_api_service, "api_root", "")
-        or "https://notificaties-api.vng.cloud/api/v1/"
+    service = config.notifications_api_service
+    assert service is not None, (
+        "Use NotificationsConfigMixin in your test case to "
+        "properly configure the notications API."
     )
+    base_url = service.api_root
     if not base_url.endswith("/"):
         base_url = f"{base_url}/"
     return base_url
