+++ src/openzaak/notifications/tests/utils.py
@@ -31,31 +31,3 @@ LOGGING_SETTINGS = {
         },
     },
 }
-
-
-class NotificationsConfigMixin:
-    @staticmethod
-    def _configure_notifications(api_root=None):
-        from notifications_api_common.models import NotificationsConfig
-        from zgw_consumers.constants import APITypes, AuthTypes
-        from zgw_consumers.models import Service
-
-        svc, _ = Service.objects.update_or_create(
-            api_root=api_root or "https://notificaties-api.vng.cloud/api/v1/",
-            defaults=dict(
-                label="NRC",
-                api_type=APITypes.nrc,
-                client_id="some-client-id",
-                secret="some-secret",
-                auth_type=AuthTypes.zgw,
-            ),
-        )
-        config = NotificationsConfig.get_solo()
-        config.notifications_api_service = svc
-        config.save()
-
-
-def get_notifications_api_root() -> str:
-    from notifications_api_common.models import NotificationsConfig
-
-    return NotificationsConfig.get_solo().notifications_api_service.api_root
